Crate vec_arena

A simple object arena.

Arena<T> is basically just a Vec<Option<T>>, which allows you to:

Insert an object (reuse an existing None element, or append to the end).

Remove object at a specified index.

Access object at a specified index.

Examples

Some data structures built using Arena<T>:

Doubly linked list

Splay tree

Structs

Arena
An object arena.

IntoIter
An iterator over the occupied slots in an Arena.

Iter
An iterator over references to the occupied slots in an Arena.

IterMut
An iterator over mutable references to the occupied slots in a Arena.